refactor: move KeycloakOpenIDConnectionManager to a separate file
