+++ src/keycloak/__init__.py
@@ -43,6 +43,7 @@ from .exceptions import (
 from .keycloak_admin import KeycloakAdmin
 from .keycloak_openid import KeycloakOpenID
 from .keycloak_uma import KeycloakUMA
+from .openid_connection import KeycloakOpenIDConnectionManager
 
 __all__ = [
     "__version__",
@@ -62,5 +63,6 @@ __all__ = [
     "KeycloakSecretNotFound",
     "KeycloakAdmin",
     "KeycloakOpenID",
+    "KeycloakOpenIDConnectionManager",
     "KeycloakUMA",
 ]
